[SQL] Do not attempt to simplify expressions that will cause panics at runtime; they may never be evaluated at runtime
