+++ sql-to-dbsp-compiler/SQL-compiler/out2983983704780668244.tmp
@@ -1,0 +1,3 @@
+
+running 95 tests
+.............................................
