+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/BaseSQLTests.java
@@ -246,7 +246,7 @@ public class BaseSQLTests {
         return new DBSPCompiler(options);
     }
 
-    static DBSPCircuit getCircuit(DBSPCompiler compiler) {
+    public static DBSPCircuit getCircuit(DBSPCompiler compiler) {
         compiler.optimize();
         String name = "circuit" + testsToRun.size();
         return compiler.getFinalCircuit(name);
