+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/ComplexQueriesTest.java
@@ -1,7 +1,6 @@
 package org.dbsp.sqlCompiler.compiler.sql;
 
 import org.dbsp.sqlCompiler.circuit.DBSPCircuit;
-import org.dbsp.sqlCompiler.circuit.operator.DBSPMapIndexOperator;
 import org.dbsp.sqlCompiler.circuit.operator.DBSPPartitionedRollingAggregateWithWaterlineOperator;
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.StderrErrorReporter;
@@ -148,149 +147,7 @@ public class ComplexQueriesTest extends BaseSQLTests {
         // this.compileRustTestCase(sql);
     }
 
-    @Test
-    public void issue1768() {
-        String sql = """
-                CREATE TABLE transaction (
-                    trans_date_time TIMESTAMP NOT NULL LATENESS INTERVAL 1 DAY,
-                    cc_num BIGINT NOT NULL,
-                    merchant STRING,
-                    category STRING,
-                    amt FLOAT64,
-                    trans_num STRING,
-                    unix_time INTEGER NOT NULL,
-                    merch_lat FLOAT64 NOT NULL,
-                    merch_long FLOAT64 NOT NULL,
-                    is_fraud INTEGER
-                );
-                
-                CREATE TABLE demographics (
-                    cc_num BIGINT NOT NULL,
-                    first STRING,
-                    gender STRING,
-                    street STRING,
-                    city STRING,
-                    state STRING,
-                    zip INTEGER,
-                    lat FLOAT64,
-                    long FLOAT64,
-                    city_pop INTEGER,
-                    job STRING,
-                    dob STRING
-                );
-            
-                CREATE VIEW V AS SELECT
-                    transaction.cc_num,
-                    CASE
-                      WHEN dayofweek(trans_date_time) IN(6, 7) THEN true
-                      ELSE false
-                    END AS is_weekend,
-                    CASE
-                      WHEN hour(trans_date_time) <= 6 THEN true
-                      ELSE false
-                    END AS is_night,
-                    category,
-                    AVG(amt) OVER window_1_day AS avg_spend_pd,
-                    AVG(amt) OVER window_7_day AS avg_spend_pw,
-                    AVG(amt) OVER window_30_day AS avg_spend_pm,
-                    COUNT(*) OVER window_1_day AS trans_freq_24,
-                      amt, state, job, unix_time, city_pop, is_fraud
-                  FROM transaction
-                  JOIN demographics
-                  ON transaction.cc_num = demographics.cc_num
-                  WINDOW
-                    window_1_day AS (PARTITION BY transaction.cc_num ORDER BY unix_time RANGE BETWEEN 86400 PRECEDING AND CURRENT ROW),
-                    window_7_day AS (PARTITION BY transaction.cc_num ORDER BY unix_time RANGE BETWEEN 604800 PRECEDING AND CURRENT ROW),
-                    window_30_day AS (PARTITION BY transaction.cc_num ORDER BY unix_time RANGE BETWEEN 2592000 PRECEDING AND CURRENT ROW);""";
-
-        DBSPCompiler compiler = this.testCompiler();
-        compiler.options.languageOptions.incrementalize = true;
-        compiler.compileStatements(sql);
-        DBSPCircuit circuit = getCircuit(compiler);
-        CircuitVisitor visitor = new CircuitVisitor(new StderrErrorReporter()) {
-            int mapIndex = 0;
-
-            @Override
-            public void postorder(DBSPMapIndexOperator operator) {
-                this.mapIndex++;
-            }
-
-            @Override
-            public void endVisit() {
-                // We expect 9 MapIndex operators instead of 11 if CSE works
-                Assert.assertEquals(9, this.mapIndex);
-            }
-        };
-        visitor.apply(circuit);
-    }
-
-    @Test
-    public void missingCast() {
-        String sql = """
-                create table TRANSACTION (unix_time BIGINT LATENESS 0);
-                """;
-        this.compileRustTestCase(sql);
-    }
-
-    @Test
-    public void issue1793() {
-        String sql = """
-                CREATE TABLE transaction_demographics (
-                    trans_date_time TIMESTAMP,
-                    cc_num BIGINT NOT NULL,
-                    category STRING,
-                    amt FLOAT64,
-                    unix_time INTEGER NOT NULL LATENESS 86400,
-                    first STRING,
-                    state STRING,
-                    job STRING,
-                    city_pop INTEGER,
-                    is_fraud BOOLEAN
-                );
-            
-                CREATE VIEW V AS SELECT
-                    cc_num,
-                    CASE
-                      WHEN dayofweek(trans_date_time) IN(6, 7) THEN true
-                      ELSE false
-                    END AS is_weekend,
-                    CASE
-                      WHEN hour(trans_date_time) <= 6 THEN true
-                      ELSE false
-                    END AS is_night,
-                    category,
-                    AVG(amt) OVER window_1_day AS avg_spend_pd,
-                    AVG(amt) OVER window_7_day AS avg_spend_pw,
-                    AVG(amt) OVER window_30_day AS avg_spend_pm,
-                    COUNT(*) OVER window_1_day AS trans_freq_24,
-                      amt, state, job, unix_time, city_pop, is_fraud
-                  FROM transaction_demographics
-                  WINDOW
-                    window_1_day AS (PARTITION BY cc_num ORDER BY unix_time RANGE BETWEEN 86400 PRECEDING AND CURRENT ROW),
-                    window_7_day AS (PARTITION BY cc_num ORDER BY unix_time RANGE BETWEEN 604800 PRECEDING AND CURRENT ROW),
-                    window_30_day AS (PARTITION BY cc_num ORDER BY unix_time RANGE BETWEEN 2592000 PRECEDING AND CURRENT ROW);""";
-        DBSPCompiler compiler = this.testCompiler();
-        compiler.options.languageOptions.incrementalize = true;
-        compiler.compileStatements(sql);
-        DBSPCircuit circuit = getCircuit(compiler);
-        CircuitVisitor visitor = new CircuitVisitor(new StderrErrorReporter()) {
-            int count = 0;
-
-            @Override
-            public void postorder(DBSPPartitionedRollingAggregateWithWaterlineOperator operator) {
-                this.count++;
-            }
-
-            @Override
-            public void endVisit() {
-                Assert.assertEquals(3, this.count);
-            }
-        };
-        visitor.apply(circuit);
-        this.compileRustTestCase(sql);
-    }
-
-    @Test @Ignore("Cross apply not yet implemented")
+   @Test @Ignore("Cross apply not yet implemented")
     public void testCrossApply() {
         String query = """
                  select d.DocumentID, ds.Status, ds.DateCreated
